rice/common/pnoc-v2: make container (re)starts more robust

+++ rice/common/pnoc-v2/clean.sh
@@ -3,10 +3,16 @@ set -u
 
 nam="$1"
 cid="$(basename "$(readlink "/run/pnoc/netns/$nam")")"
+if [ -z "$cid" ]; then
+	echo "No such container '$nam'!" >&2
+	exit
+fi
 
+# remove container registration
 rm -f "/run/pnoc/netns/$nam" "/run/pnoc/hosts/$nam"
 ip netns del "$cid"
 
+# reset port forward
 ports="/etc/pnoc/ports/$nam"
 if [ -e "$ports" ]; then
 	while read -r ipv typ src _; do
@@ -14,11 +20,12 @@ if [ -e "$ports" ]; then
 		4) tbl=ip ;;
 		6) tbl=ip6 ;;
 		*)
-			echo "port forward: invalid IP type $ipv!" >&2
+			echo "port forward: invalid IP type '$ipv'!" >&2
 			continue
 			;;
 		esac
-		nft delete element "$tbl" io.systemd.nat map_port_ipport "{ $typ . $src }"
+
+		nft delete element "$tbl" io.systemd.nat map_port_ipport "{ $typ . $src }" 2>/dev/null
 	done <"$ports"
 fi
